requirements.txt
discord.py[voice]
asyncpg[speedups]
aiohttp
jishaku
psutil
pylance